requirements.txt
requests==2.28.0
pyTelegramBotAPI==4.6.0
mysql-connector==2.2.9
pyzbar==0.1.9
Pillow==9.1.1
ptvsd==4.3.2 # Required for debugging.
